+++ resources/views/hour/create.blade.php
@@ -20,6 +20,8 @@
             <x-laraknife.forms.string type="number" position="first" name="factor" label="Factor"
                 value="{{ $context->valueOf('factor') }}" width2="4" />
             <x-laraknife.forms.combobox position="last" name="owner_id" label="Owner" :options="$optionsOwner" width2="4" />
+            <x-laraknife.forms.string type="number" position="alone" name="interested" label="Interested"
+                value="{{ $context->valueOf('interested') }}" width2="4" />
             <x-laraknife.forms.text position="alone" name="description" label="Description"
                 value="{{ $context->valueOf('description') }}" width2="10" rows="2" />
         </x-laraknife.panels.create>